+++ src/cpp11.cpp
@@ -33,10 +33,10 @@ extern "C" SEXP _bignum_c_bigfloat_to_double(SEXP x) {
   END_CPP11
 }
 // bigfloat_interface.cpp
-cpp11::strings c_bigfloat_format(cpp11::strings x);
-extern "C" SEXP _bignum_c_bigfloat_format(SEXP x) {
+cpp11::strings c_bigfloat_format(cpp11::strings x, cpp11::strings notation, cpp11::integers digits, bool is_sigfig);
+extern "C" SEXP _bignum_c_bigfloat_format(SEXP x, SEXP notation, SEXP digits, SEXP is_sigfig) {
   BEGIN_CPP11
-    return cpp11::as_sexp(c_bigfloat_format(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(x)));
+    return cpp11::as_sexp(c_bigfloat_format(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(x), cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(notation), cpp11::as_cpp<cpp11::decay_t<cpp11::integers>>(digits), cpp11::as_cpp<cpp11::decay_t<bool>>(is_sigfig)));
   END_CPP11
 }
 // bigfloat_interface.cpp
@@ -319,6 +319,20 @@ extern "C" SEXP _bignum_c_bigfloat_lgamma(SEXP lhs) {
     return cpp11::as_sexp(c_bigfloat_lgamma(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(lhs)));
   END_CPP11
 }
+// bigfloat_interface.cpp
+cpp11::strings c_bigfloat_digamma(cpp11::strings lhs);
+extern "C" SEXP _bignum_c_bigfloat_digamma(SEXP lhs) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(c_bigfloat_digamma(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(lhs)));
+  END_CPP11
+}
+// bigfloat_interface.cpp
+cpp11::strings c_bigfloat_trigamma(cpp11::strings lhs);
+extern "C" SEXP _bignum_c_bigfloat_trigamma(SEXP lhs) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(c_bigfloat_trigamma(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(lhs)));
+  END_CPP11
+}
 // biginteger_interface.cpp
 cpp11::strings c_biginteger(cpp11::strings x);
 extern "C" SEXP _bignum_c_biginteger(SEXP x) {
@@ -348,10 +362,10 @@ extern "C" SEXP _bignum_c_biginteger_to_double(SEXP x) {
   END_CPP11
 }
 // biginteger_interface.cpp
-cpp11::strings c_biginteger_format(cpp11::strings x);
-extern "C" SEXP _bignum_c_biginteger_format(SEXP x) {
+cpp11::strings c_biginteger_format(cpp11::strings x, cpp11::strings notation);
+extern "C" SEXP _bignum_c_biginteger_format(SEXP x, SEXP notation) {
   BEGIN_CPP11
-    return cpp11::as_sexp(c_biginteger_format(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(x)));
+    return cpp11::as_sexp(c_biginteger_format(cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(x), cpp11::as_cpp<cpp11::decay_t<cpp11::strings>>(notation)));
   END_CPP11
 }
 // biginteger_interface.cpp
@@ -486,11 +500,12 @@ extern SEXP _bignum_c_bigfloat_cummax(SEXP);
 extern SEXP _bignum_c_bigfloat_cummin(SEXP);
 extern SEXP _bignum_c_bigfloat_cumprod(SEXP);
 extern SEXP _bignum_c_bigfloat_cumsum(SEXP);
+extern SEXP _bignum_c_bigfloat_digamma(SEXP);
 extern SEXP _bignum_c_bigfloat_divide(SEXP, SEXP);
 extern SEXP _bignum_c_bigfloat_exp(SEXP);
 extern SEXP _bignum_c_bigfloat_expm1(SEXP);
 extern SEXP _bignum_c_bigfloat_floor(SEXP);
-extern SEXP _bignum_c_bigfloat_format(SEXP);
+extern SEXP _bignum_c_bigfloat_format(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _bignum_c_bigfloat_gamma(SEXP);
 extern SEXP _bignum_c_bigfloat_lgamma(SEXP);
 extern SEXP _bignum_c_bigfloat_log(SEXP);
@@ -513,6 +528,7 @@ extern SEXP _bignum_c_bigfloat_tanh(SEXP);
 extern SEXP _bignum_c_bigfloat_to_double(SEXP);
 extern SEXP _bignum_c_bigfloat_to_integer(SEXP);
 extern SEXP _bignum_c_bigfloat_to_logical(SEXP);
+extern SEXP _bignum_c_bigfloat_trigamma(SEXP);
 extern SEXP _bignum_c_bigfloat_trunc(SEXP);
 extern SEXP _bignum_c_biginteger(SEXP);
 extern SEXP _bignum_c_biginteger_abs(SEXP);
@@ -522,7 +538,7 @@ extern SEXP _bignum_c_biginteger_cummax(SEXP);
 extern SEXP _bignum_c_biginteger_cummin(SEXP);
 extern SEXP _bignum_c_biginteger_cumprod(SEXP);
 extern SEXP _bignum_c_biginteger_cumsum(SEXP);
-extern SEXP _bignum_c_biginteger_format(SEXP);
+extern SEXP _bignum_c_biginteger_format(SEXP, SEXP);
 extern SEXP _bignum_c_biginteger_modulo(SEXP, SEXP);
 extern SEXP _bignum_c_biginteger_multiply(SEXP, SEXP);
 extern SEXP _bignum_c_biginteger_pow(SEXP, SEXP);
@@ -554,11 +570,12 @@ static const R_CallMethodDef CallEntries[] = {
     {"_bignum_c_bigfloat_cummin",       (DL_FUNC) &_bignum_c_bigfloat_cummin,       1},
     {"_bignum_c_bigfloat_cumprod",      (DL_FUNC) &_bignum_c_bigfloat_cumprod,      1},
     {"_bignum_c_bigfloat_cumsum",       (DL_FUNC) &_bignum_c_bigfloat_cumsum,       1},
+    {"_bignum_c_bigfloat_digamma",      (DL_FUNC) &_bignum_c_bigfloat_digamma,      1},
     {"_bignum_c_bigfloat_divide",       (DL_FUNC) &_bignum_c_bigfloat_divide,       2},
     {"_bignum_c_bigfloat_exp",          (DL_FUNC) &_bignum_c_bigfloat_exp,          1},
     {"_bignum_c_bigfloat_expm1",        (DL_FUNC) &_bignum_c_bigfloat_expm1,        1},
     {"_bignum_c_bigfloat_floor",        (DL_FUNC) &_bignum_c_bigfloat_floor,        1},
-    {"_bignum_c_bigfloat_format",       (DL_FUNC) &_bignum_c_bigfloat_format,       1},
+    {"_bignum_c_bigfloat_format",       (DL_FUNC) &_bignum_c_bigfloat_format,       4},
     {"_bignum_c_bigfloat_gamma",        (DL_FUNC) &_bignum_c_bigfloat_gamma,        1},
     {"_bignum_c_bigfloat_lgamma",       (DL_FUNC) &_bignum_c_bigfloat_lgamma,       1},
     {"_bignum_c_bigfloat_log",          (DL_FUNC) &_bignum_c_bigfloat_log,          1},
@@ -581,6 +598,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_bignum_c_bigfloat_to_double",    (DL_FUNC) &_bignum_c_bigfloat_to_double,    1},
     {"_bignum_c_bigfloat_to_integer",   (DL_FUNC) &_bignum_c_bigfloat_to_integer,   1},
     {"_bignum_c_bigfloat_to_logical",   (DL_FUNC) &_bignum_c_bigfloat_to_logical,   1},
+    {"_bignum_c_bigfloat_trigamma",     (DL_FUNC) &_bignum_c_bigfloat_trigamma,     1},
     {"_bignum_c_bigfloat_trunc",        (DL_FUNC) &_bignum_c_bigfloat_trunc,        1},
     {"_bignum_c_biginteger",            (DL_FUNC) &_bignum_c_biginteger,            1},
     {"_bignum_c_biginteger_abs",        (DL_FUNC) &_bignum_c_biginteger_abs,        1},
@@ -590,7 +608,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_bignum_c_biginteger_cummin",     (DL_FUNC) &_bignum_c_biginteger_cummin,     1},
     {"_bignum_c_biginteger_cumprod",    (DL_FUNC) &_bignum_c_biginteger_cumprod,    1},
     {"_bignum_c_biginteger_cumsum",     (DL_FUNC) &_bignum_c_biginteger_cumsum,     1},
-    {"_bignum_c_biginteger_format",     (DL_FUNC) &_bignum_c_biginteger_format,     1},
+    {"_bignum_c_biginteger_format",     (DL_FUNC) &_bignum_c_biginteger_format,     2},
     {"_bignum_c_biginteger_modulo",     (DL_FUNC) &_bignum_c_biginteger_modulo,     2},
     {"_bignum_c_biginteger_multiply",   (DL_FUNC) &_bignum_c_biginteger_multiply,   2},
     {"_bignum_c_biginteger_pow",        (DL_FUNC) &_bignum_c_biginteger_pow,        2},